refactor: extract Alternatives component from Alternatives controller
