+++ pix-editor/app/components/alternatives.js
@@ -1,0 +1,15 @@
+import Component from '@glimmer/component';
+import { inject as service } from '@ember/service';
+import { tracked } from '@glimmer/tracking';
+
+export default class Alternatives extends Component {
+
+  @tracked competence = null;
+  @tracked arePerimeDeclisDisplayed = false;
+
+  @service config;
+
+  get alternatives() {
+    return this.arePerimeDeclisDisplayed ? this.args.challenge.alternatives : this.args.challenge.alternatives.filter(alternative => alternative.status !== 'périmé');
+  }
+}
